refactor: move dead code widgets into page feature
