+++ src/pages/dead-code/PageDeadCode.tsx
@@ -1,4 +1,4 @@
-import { DeadCodePanel } from './widgets/DeadCodePanel/DeadCodePanel';
+import { DeadCodePanel } from './features/DeadCodePanel/ui/DeadCodePanel';
 
 export function PageDeadCode() {
   return <DeadCodePanel />;
