+++ src/pages/dead-code/features/DeadCodePanel/ui/DeadCodeCategoryHeader.tsx
