+++ src/pages/dead-code/features/DeadCodePanel/ui/DeadCodePanel.tsx
@@ -5,7 +5,7 @@
 
 import { useAtomValue } from 'jotai';
 import { Sparkles } from 'lucide-react';
-import { useRef, useState } from 'react';
+import { useState } from 'react';
 import { useDeadCodeAnalysis } from '@/features/Code/CodeAnalyzer/DeadCodeAnalyzer/lib/useDeadCodeAnalysis.ts';
 import {
   deadCodeResultsAtom,
@@ -15,16 +15,15 @@ import IDEScrollView from '@/features/File/OpenFiles/ui/IDEScrollView/IDEScrollV
 import { RefactoringPromptDialog } from '@/features/RefactoringPrompt/RefactoringPromptDialog.tsx';
 import { Button } from '@/shared/ui/Button';
 import { Sidebar } from '@/shared/ui/Sidebar';
-import { DeadCodeExplorer } from '../DeadCodeExplorer/DeadCodeExplorer.tsx';
-import { DeadCodePanelHeader } from './ui/DeadCodePanelHeader.tsx';
-import { DeadCodePanelSummary } from './ui/DeadCodePanelSummary.tsx';
+import { DeadCodeExplorer } from './DeadCodeExplorer.tsx';
+import { DeadCodePanelHeader } from './DeadCodePanelHeader.tsx';
+import { DeadCodePanelSummary } from './DeadCodePanelSummary.tsx';
 
-export function DeadCodePanel({ className: _className }: { className?: string }) {
+export function DeadCodePanel() {
   useDeadCodeAnalysis(); // Auto-analyze on mount
   const deadCodeResults = useAtomValue(deadCodeResultsAtom);
   const selectedItems = useAtomValue(selectedDeadCodeItemsAtom);
   const [showPromptDialog, setShowPromptDialog] = useState(false);
-  const containerRef = useRef<HTMLDivElement>(null);
 
   function handleGeneratePromptClick() {
     setShowPromptDialog(true);
@@ -37,14 +36,14 @@ export function DeadCodePanel({ className: _className }: { className?: string })
   return (
     <div className="flex h-full w-full overflow-hidden">
       {/* 좌측: Dead Code Panel */}
-      <div ref={containerRef} className="relative focus:outline-none">
+      <div className="relative focus:outline-none">
         <Sidebar resizable defaultWidth={280} minWidth={200} maxWidth={600} className="h-full shadow-2xl">
           <Sidebar.Header>
             <DeadCodePanelHeader />
           </Sidebar.Header>
 
           <DeadCodePanelSummary />
-          <DeadCodeExplorer containerRef={containerRef} />
+          <DeadCodeExplorer />
 
           {/* Generate Prompt Button */}
           {deadCodeResults && selectedItems.size > 0 && (
