+++ src/pages/dead-code/features/DeadCodePanel/ui/DeadCodePanelHeader.tsx
