+++ src/pages/dead-code/features/DeadCodePanel/ui/DeadCodePanelSummary.tsx
