+++ src/pages/dead-code/features/DeadCodePanel/ui/DeadCodeResultItem.tsx
